Some refactoring.
Extract some code into separate methods: ProcessFile and ProcessText (mainly for testability later).

If the text being read contains Unicode replacement character (code point 65533 '\uFFFD' represented as 0xEF 0xBF 0xBD in UTF-8), try opening the file with Encoding.Default instead. This way we support UTF-8 (with or without BOM), UTF16 (LE or BE) and if we're lucky a non-unicode (ANSI?) encoding as well.