Change RemoveConsecutiveEmptyLines for C# files to be based on Roslyn instead of blind text replace.
The problem was that it was removing empty lines even from multiline string literals. That's bad; we don't want to change multiline string literals ever. Reliably determining whether we're inside of a multiline string literal would mean reimplementing half of the C# scanner. So it's easier to switch to using Roslyn to implement this fix.

Now we remove consecutive empty lines as part of FormatSolution (that is already based on Roslyn), and keep the simple implementation for non-C# files (ps1, psm1).